Added member function template `of<T>()` to min_t and max_t to be more succinct than a static_cast. Added _shift_v and _width_v constants to BITFIELD Added knuth_lcg Added default (double) template arguments to constants kblib::pi &c.
Changed comparison operators of min_t and max_t to be hidden friends.
Silenced warnings in test case in sort.cpp

+++ sort.cpp
@@ -164,8 +164,9 @@ TEST_CASE("insertion sort performance") {
 		};
 
 		TIME(
-		    time_per, 30, 1000,
-		    +[](std::size_t i) { return static_cast<double>(i) * i; });
+		    time_per, 30, 1000, +[](std::size_t i) {
+			    return static_cast<double>(i) * static_cast<double>(i);
+		    });
 	}
 	SECTION("adaptive_insertion_sort_copy on sorted data is fast") {
 		auto time_per = [](std::size_t size) {
